+++ app/decorators/controllers/account_requests_controller_decorator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-PafsCore::AccountRequestsController.class_eval do
+AccountRequestsController.class_eval do
   helper ApplicationHelper
   layout ->(_) { "pafs" }
 end
